+++ crates/pertylizer/src/patches/chaos_drone.rs
@@ -91,7 +91,7 @@ different due to the chaotic nature of the algorithm.
     patch.add_module(
         ModuleBuilder::new(1, ModuleType::Amplifier)
             .position(850.0, 50.0)
-            .param_f("level", 0.6)
+            .param_f("level", 1.5)
             .build(),
     );
 
@@ -118,7 +118,7 @@ different due to the chaotic nature of the algorithm.
     patch.add_module(
         ModuleBuilder::new(1, ModuleType::StereoOutput)
             .position(1200.0, 50.0)
-            .param_f("master", 0.7)
+            .param_f("master", 0.9)
             .build(),
     );